server/requirements.txt
reportbro-lib>=3.11.2
Flask>=3.1.2
Flask-Cors>=6.0.1